Fix #6855: Non-RST translated text should be parsed by the appropriate parser
